HV-798 Creating dedicated package for message descriptor parser
