+++ engine/src/main/java/org/hibernate/validator/internal/engine/messageinterpolation/parser/MessageState.java
@@ -14,36 +14,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hibernate.validator.internal.engine.messageinterpolation;
+package org.hibernate.validator.internal.engine.messageinterpolation.parser;
 
+import org.hibernate.validator.internal.engine.messageinterpolation.InterpolationTermType;
 import org.hibernate.validator.internal.util.logging.Log;
 import org.hibernate.validator.internal.util.logging.LoggerFactory;
 
 /**
  * @author Hardy Ferentschik
  */
-public class BeginState implements ParserState {
+public class MessageState implements ParserState {
 	private static final Log log = LoggerFactory.make();
 
 	@Override
-	public void terminate(ParserContext context) throws MessageDescriptorFormatException {
+	public void start(ParserContext context) {
+		throw new IllegalStateException( "The parsing of the message descriptor cannot start in this state." );
 	}
 
 	@Override
-	public void start(ParserContext context) throws MessageDescriptorFormatException {
-		context.next();
+	public void terminate(ParserContext context) throws MessageDescriptorFormatException {
+		context.terminateToken();
 	}
 
 	@Override
 	public void handleNonMetaCharacter(char character, ParserContext context)
 			throws MessageDescriptorFormatException {
 		context.appendToToken( character );
-		context.transitionState( new MessageState() );
 		context.next();
 	}
 
 	@Override
 	public void handleBeginTerm(char character, ParserContext context) throws MessageDescriptorFormatException {
+		context.terminateToken();
+
 		context.appendToToken( character );
 		if ( context.getInterpolationType().equals( InterpolationTermType.PARAMETER ) ) {
 			context.makeParameterToken();
@@ -54,13 +57,17 @@ public class BeginState implements ParserState {
 
 	@Override
 	public void handleEndTerm(char character, ParserContext context) throws MessageDescriptorFormatException {
-		throw log.getNonTerminatedParameterException( context.getOriginalMessageDescriptor(), character );
+		throw log.getNonTerminatedParameterException(
+				context.getOriginalMessageDescriptor(),
+				character
+		);
 	}
 
 	@Override
 	public void handleEscapeCharacter(char character, ParserContext context)
 			throws MessageDescriptorFormatException {
 		context.appendToToken( character );
+
 		context.transitionState( new EscapedState( this ) );
 		context.next();
 	}
@@ -71,8 +78,7 @@ public class BeginState implements ParserState {
 			handleNonMetaCharacter( character, context );
 		}
 		else {
-			ParserState state = new ELState();
-			context.transitionState( state );
+			context.transitionState( new ELState() );
 			context.next();
 		}
 	}
