+++ engine/src/main/java/org/hibernate/validator/internal/engine/messageinterpolation/parser/Token.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hibernate.validator.internal.engine.messageinterpolation;
+package org.hibernate.validator.internal.engine.messageinterpolation.parser;
 
 /**
  * Parsing a message descriptor results in multiple {@code Token}s. Each token
