+++ engine/src/main/java/org/hibernate/validator/internal/util/logging/Log.java
@@ -38,7 +38,7 @@ import javax.validation.ValidationException;
 import javax.xml.bind.JAXBException;
 import javax.xml.stream.XMLStreamException;
 
-import org.hibernate.validator.internal.engine.messageinterpolation.MessageDescriptorFormatException;
+import org.hibernate.validator.internal.engine.messageinterpolation.parser.MessageDescriptorFormatException;
 import org.jboss.logging.BasicLogger;
 import org.jboss.logging.Cause;
 import org.jboss.logging.LogMessage;
